feat(apps/web): extract SongCard component from album view
Extract the song card UI into a reusable SongCard component and refactor
the album tracks view to use it. This removes duplicate markup and the
formatDuration utility from the album route, improving maintainability
and enabling reuse across other parts of the application.
